added customization.h to centralize all the customization string
git-svn-id: https://outlook-client.forge.funambol.org/svn/windows-client/trunk@169 31df160e-6edc-4e4b-9207-171dac44fd17

+++ windows/trunk/outlook/mainclientDll/src/cpp/outlook/ClientContact.cpp
@@ -493,7 +493,7 @@ const wstring ClientContact::getComplexProperty(const wstring& propertyName) {
                 if (separatorEnd == StringBuffer::npos){
                     separatorEnd = valueBuffer.length();
                 }
-                if (separatorEnd < valueBuffer.length()){
+                if (separatorEnd < (int)(valueBuffer.length())){
                     address1.append(valueBuffer.substr(separatorEnd +1, (valueBuffer.length() - separatorEnd -1)));
                 }
                 address2 = valueBuffer.substr((separator + extAdrSeparatorLenght), (separatorEnd - separator - extAdrSeparatorLenght) );
